+++ apps/native-component-list/src/screens/AppleAuthenticationScreen.tsx
@@ -32,9 +32,7 @@ const CREDENTIAL_MESSAGES = {
   [AppleAuthenticationCredentialState.TRANSFERRED]: 'Credentials transferred.', // Whatever that means...
 };
 
-// See: https://github.com/expo/expo/pull/10229#discussion_r490961694
-// eslint-disable-next-line @typescript-eslint/ban-types
-export default class AppleAuthenticationScreen extends React.Component<{}, State> {
+export default class AppleAuthenticationScreen extends React.Component<object, State> {
   static navigationOptions = {
     title: 'Apple Authentication',
   };