+++ apps/router-e2e/__e2e__/css-global-import/app/index.js
@@ -1,0 +1,16 @@
+import { Text } from 'react-native';
+
+import '../global.css';
+
+export default function Page() {
+  return (
+    <>
+      <Text testID="index-text" style={{ $$css: true, _: 'collide' }}>
+        Index
+      </Text>
+      <Text testID="beta-text" style={{ $$css: true, _: 'beta' }}>
+        Index
+      </Text>
+    </>
+  );
+}