+++ packages/@expo/cli/e2e/__tests__/export.test.ts
@@ -93,7 +93,7 @@ describe('server', () => {
                 path: 'assets/3858f62230ac3c915f300c664312c63f',
               },
             ],
-            bundle: expect.stringMatching(/_expo\/static\/js\/android\/AppEntry-.*\.hbc/),
+            bundle: expect.stringMatching(/_expo\/static\/js\/android\/AppEntry-.*\.js/),
           },
           ios: {
             assets: [
@@ -110,7 +110,7 @@ describe('server', () => {
                 path: 'assets/2f334f6c7ca5b2a504bdf8acdee104f3',
               },
             ],
-            bundle: expect.stringMatching(/_expo\/static\/js\/ios\/AppEntry-.*\.hbc/),
+            bundle: expect.stringMatching(/_expo\/static\/js\/ios\/AppEntry-.*\.js/),
           },
         },
         version: 0,
@@ -161,10 +161,10 @@ describe('server', () => {
 
       // If this changes then everything else probably changed as well.
       expect(files).toEqual([
-        expect.stringMatching(/_expo\/static\/js\/android\/AppEntry-[\w\d]+\.hbc/),
-        expect.stringMatching(/_expo\/static\/js\/android\/AppEntry-[\w\d]+\.hbc\.map/),
-        expect.stringMatching(/_expo\/static\/js\/ios\/AppEntry-[\w\d]+\.hbc/),
-        expect.stringMatching(/_expo\/static\/js\/ios\/AppEntry-[\w\d]+\.hbc\.map/),
+        expect.stringMatching(/_expo\/static\/js\/android\/AppEntry-[\w\d]+\.js/),
+        expect.stringMatching(/_expo\/static\/js\/android\/AppEntry-[\w\d]+\.js\.map/),
+        expect.stringMatching(/_expo\/static\/js\/ios\/AppEntry-[\w\d]+\.js/),
+        expect.stringMatching(/_expo\/static\/js\/ios\/AppEntry-[\w\d]+\.js\.map/),
         expect.stringMatching(/_expo\/static\/js\/web\/AppEntry-[\w\d]+\.js/),
         expect.stringMatching(/_expo\/static\/js\/web\/AppEntry-[\w\d]+\.js\.map/),
         'assetmap.json',
@@ -176,7 +176,6 @@ describe('server', () => {
         'assets/assets/icon.8034d8318b239108719ff3f22f31ef15@2x.png',
 
         'assets/fb960eb5e4eb49ec8786c7f6c4a57ce2',
-        'debug.html',
         'favicon.ico',
         'index.html',
         'metadata.json',
@@ -282,7 +281,6 @@ describe('server', () => {
         'assets/2f334f6c7ca5b2a504bdf8acdee104f3',
         'assets/9ce7db807e4147e00df372d053c154c2',
         'assets/fb960eb5e4eb49ec8786c7f6c4a57ce2',
-        'debug.html',
         'favicon.ico',
         'metadata.json',
       ]);