+++ packages/expo-image-picker/android/src/main/java/expo/modules/imagepicker/contracts/CameraContract.kt
@@ -30,10 +30,10 @@ internal class CameraContract(
     }.contentResolver
 
   override fun createIntent(context: Context, input: CameraContractOptions): Intent =
-    Intent(input.options.mediaTypes.toCameraIntentAction())
+    Intent(input.options.nativeMediaTypes.toCameraIntentAction())
       .putExtra(MediaStore.EXTRA_OUTPUT, input.uri.toUri())
       .apply {
-        if (input.options.mediaTypes.toCameraIntentAction() == MediaStore.ACTION_VIDEO_CAPTURE) {
+        if (input.options.nativeMediaTypes.toCameraIntentAction() == MediaStore.ACTION_VIDEO_CAPTURE) {
           putExtra(MediaStore.EXTRA_DURATION_LIMIT, input.options.videoMaxDuration)
         }
         if (input.options.cameraType == CameraType.FRONT) {