+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/exception/CodedException.kt
@@ -27,6 +27,7 @@ open class CodedException(
   // the javaClass property in the constructor.
   private var providedCode: String? = null
 
+  @get:DoNotStrip
   val code
     get() = providedCode ?: inferCode(javaClass)
 
@@ -185,6 +186,13 @@ internal class InvalidSharedObjectException(
   sharedType: KType
 ) : CodedException(
   message = "Cannot convert provided JavaScriptObject to the '$sharedType', because it doesn't contain valid id"
+)
+
+internal class IncorrectRefTypeException(
+  desiredType: KType,
+  receivedClass: Class<*>
+) : CodedException(
+  message = "Cannot convert received '$receivedClass' to the '$desiredType', because of the inner ref type mismatch"
 )
 
 internal class FieldCastException(