+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/types/Either.kt
@@ -33,6 +33,14 @@ class UnconvertedValue(
 
 data class ConvertedValue(val convertedValue: Any) : DeferredValue()
 
+/**
+ * Converts the type T to KClass. It preserves all parameters of type T if it's a generic.
+ * For instance, `toKClass<List<String>>() == KClass<List<String>>`
+ * Getting `KClass<List<String>>` isn't possible without a helper function because
+ * `List<String>::class` isn't a valid syntax.
+ */
+inline fun <reified T : Any> toKClass(): KClass<T> = T::class
+
 @EitherType
 @DoNotStrip
 // We can't strip this class because typeOf<Either<T,P>> won't work in the release builds.
@@ -102,9 +110,9 @@ open class EitherOfThree<FirstType : Any, SecondType : Any, ThirdType : Any>(
   fun `is`(@Suppress("UNUSED_PARAMETER") type: KClass<ThirdType>): Boolean = `is`(2)
 
   @JvmName("getThirdType")
-  fun get(@Suppress("UNUSED_PARAMETER") type: KClass<ThirdType>) = get(3) as ThirdType
+  fun get(@Suppress("UNUSED_PARAMETER") type: KClass<ThirdType>) = get(2) as ThirdType
 
-  fun third(): ThirdType = get(3) as ThirdType
+  fun third(): ThirdType = get(2) as ThirdType
 }
 
 @EitherType