+++ packages/expo-sqlite/src/hooks.tsx
@@ -131,7 +131,7 @@ export function SQLiteProvider({
  *
  * export function Main() {
  *   const db = useSQLiteContext();
- *   console.log('sqlite version', db.getSync('SELECT sqlite_version()'));
+ *   console.log('sqlite version', db.getFirstSync('SELECT sqlite_version()'));
  *   return <View />
  * }
  * ```